requirements.txt
Django>=1.3,<1.8
south
whoosh
ipy
python-ldap
djangorestframework>=2.0.0,<3.0.0
